+++ jdaviz/configs/imviz/tests/test_parser.py
@@ -282,9 +282,9 @@ class TestParseImage:
         assert_allclose(sky.dec.deg, -69.49460838)
         data_unit = u.MJy / u.sr
         assert_quantity_allclose(tbl[0]['background'], 0.1741226315498352 * data_unit)
-        assert_quantity_allclose(tbl[0]['npix'], 111.22023392 * u.pix)
-        assert_quantity_allclose(tbl[0]['aperture_sum'], 4.989882e-09 * u.MJy)
-        assert_quantity_allclose(tbl[0]['pixarea_tot'], 1.0384377922763469e-11 * u.sr)
+        assert_quantity_allclose(tbl[0]['npix'], 109 * u.pix)
+        assert_quantity_allclose(tbl[0]['aperture_sum'], 4.890271e-09 * u.MJy)
+        assert_quantity_allclose(tbl[0]['pixarea_tot'], 1.017708e-11 * u.sr)
         assert_quantity_allclose(tbl[0]['aperture_sum_counts'], 132061.576643 * u.count)
         assert_quantity_allclose(tbl[0]['aperture_sum_counts_err'], 363.402775 * u.count)
         assert_quantity_allclose(tbl[0]['counts_fac'], 0.0036385915646798953 * (data_unit / u.ct))